+++ src/hooks/index.js
@@ -1,5 +1,6 @@
 export { default as useBoundWithTitlesByHrids } from './useBoundWithTitlesByHrids';
 export { default as useBrowseValidation } from './useBrowseValidation';
+export { default as useConsortiumItems } from './useConsortiumItems';
 export { default as useCallout } from './useCallout';
 export { default as useHoldingItemsQuery } from './useHoldingItemsQuery';
 export { default as useHoldingMutation } from './useHoldingMutation';